+++ part2/courseinfo/src/App.js
@@ -1,64 +1,60 @@
-const Header = props => <h1>{props.name}</h1>
-
-const Content = ({part}) => {
-  const name = part.name
-  const exercises = part.exercises
-  const id = part.id
-
-  return <div>{name} {exercises}</div>
-}
-
-const Course = ({course}) => {
-  const name = course.name
-  const parts = course.parts
-  let sum = 0
-  let array1 = []
-  
-  // for (let i = 0; i < parts.length; i++) {
-  //   sum += parts[i].exercises
-  // } // works fine btw
-
-  for (let i = 0; i < parts.length; i++) {
-    array1.push(parts[i].exercises)
-  }
-
-  const sumWithInitial = array1.reduce(
-  (previousValue, currentValue) => previousValue + currentValue,
-  0);
-
-  return <div>
-    <Header name={name}/>
-    <Content part={parts[0]} />
-    <Content part={parts[1]} />
-    <Content part={parts[2]} />
-    <div>sum of exercises: {sumWithInitial}</div>
-  </div>
-}
+import Course from './components/Course.js'
+// import Header from './components/Header.js'
 
 const App = () => {
-  const course = {
-    id: 1,
-    name: 'Half Stack application development',
-    parts: [
-      {
-        name: 'Fundamentals of React',
-        exercises: 10,
-        id: 1
-      },
-      {
-        name: 'Using props to pass data',
-        exercises: 7,
-        id: 2
-      },
-      {
-        name: 'State of a component',
-        exercises: 14,
-        id: 3
-      }
-    ]
-  }
+  const courses = [
+    {
+      name: 'Half Stack application development',
+      id: 1,
+      parts: [
+        {
+          name: 'Fundamentals of React',
+          exercises: 10,
+          id: 1
+        },
+        {
+          name: 'Using props to pass data',
+          exercises: 7,
+          id: 2
+        },
+        {
+          name: 'State of a component',
+          exercises: 14,
+          id: 3
+        },
+        {
+          name: 'Redux',
+          exercises: 11,
+          id: 4
+        }
+      ]
+    }, 
+    {
+      name: 'Node.js',
+      id: 2,
+      parts: [
+        {
+          name: 'Routing',
+          exercises: 3,
+          id: 1
+        },
+        {
+          name: 'Middlewares',
+          exercises: 7,
+          id: 2
+        }
+      ]
+    }
+  ]
 
-  return <Course course={course} />
+  console.log(courses[1])
+
+  return (
+    <div>
+      <Course course={courses[0]} />
+      <Course course={courses[1]} />
+    </div>
+  )
 }
 
 export default App
