+++ part2/courseinfo/src/components/Content.js
@@ -1,0 +1,9 @@
+const Content = ({part}) => {
+    const name = part.name
+    const exercises = part.exercises
+    // const id = part.id
+  
+    return <div>{name} {exercises}</div>
+}
+
+export default Content
