+++ part2/courseinfo/src/components/Header.js
@@ -1,0 +1,3 @@
+const Header = props => <h1>{props.name}</h1>
+
+export default Header
